+++ frontend/src/scss/components/Map.scss
@@ -33,7 +33,10 @@ main {
       gap: 1rem;
       justify-content: flex-start;
     }
-
+    @media screen and (max-width: 390px) {
+      width: 100%;
+      height: 100%;
+    }
     .mapButton {
       background-color: $color-grey;
       display: flex;
@@ -62,6 +65,9 @@ main {
         height: 50%;
         max-height: 500px;
         max-width: 90%;
+      }
+      @media screen and (max-width: 390px) {
+        display: none;
       }
       button {
         width: 7rem;
@@ -120,6 +126,11 @@ main {
         display: flex;
         width: 90vw;
         height: 50vh;
+      }
+      @media screen and (max-width: 390px) {
+        display: flex;
+        width: 90vw;
+        height: 90vh;
       }
     }
   }
